requirements.txt
asgiref==3.5.2
Django==4.1.2
django-cors-headers==3.14.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Markdown==3.4.1
Pillow==9.4.0
PyJWT==2.9.0
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.5
